* main: (21 commits)
  修正註冊的經緯度功能
  預約通知排除報到成功通知
  畫面修正
  編輯診所基本資料可抓取經緯度並預覽地圖
  預約紀錄排列順序更新
  編輯診所營業時間
  編輯診所基本資料(圖片完成)
  編輯診所基本資訊(照片未完成)
  我的收藏功能完成
  診所註冊新增經緯度功能
  編輯診所專科(僅後端)
  修改衝突
  調整醫師資料畫面呈現
  報到功能更新
  增加右側menu跳轉功能
  完成修改預約功能
  診所管理完成顯示診所資訊
  修改錯誤
  錯誤修正
  診間叫號功能串叫號機
  ...

+++ ires-system/src/main/java/web/clinic/entity/Clinic.java
@@ -1,6 +1,5 @@
 package web.clinic.entity;
 
-import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 import javax.persistence.Column;
@@ -102,13 +101,13 @@ public class Clinic extends Core{
     private Timestamp updateTime;
 
     @Column(name = "latitude")
-    private BigDecimal latitude;
+    private Double latitude;
 
     @Column(name = "longitude")
-    private BigDecimal longitude;
+    private Double longitude;
     
     @Column(name = "rating")
-    private BigDecimal rating;
+    private Double rating;
 
     @Lob
     @Column(name = "profile_picture")
